# doc/api-doc.yml
swagger: "2.0"
info:
  title: "Qlik Core QIX session service"
  description: "REST API for handling Qlik Associative Engine sessions in Qlik Core"
  version: "0.0.1"
host: "localhost:9455"
schemes:
  - "http"
basePath: "/v1"
produces:
  - "application/json"
consumes:
  - "application/json"
tags:
  - name: "session"
    description: "API for Qlik Associative Engine session handling"
  - name: "health"
    description: "Health status of the service"
paths:
  /session/doc/{docId}:
   get:
      tags:
        - "session"
      summary: "Get a session for a document with docId"
      description: "Finds a Qlik Associative Engine and opens the document for the user"
      operationId: "openSession"
      produces:
        - "application/json"
      parameters:
        - name: "docId"
          in: "path"
          description: "Guid of the doc to open"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/session"
        "404":
          description: "Document not found"
  /session/doc/{docId}/flush:
   get:
      tags:
        - "session"
      summary: "Get a session for a document with docId"
      description: "Finds a Qlik Associative Engine and opens the document for the user"
      operationId: "openSession"
      produces:
        - "application/json"
      parameters:
        - name: "docId"
          in: "path"
          description: "Guid of the doc to open"
          required: true
          type: "string"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/session"
        "404":
          description: "Document not found"
  /session/session-doc:
    get:
      tags:
        - "session"
      summary: "Get a session for a session document"
      description: "Finds a Qlik Associative Engine and creates a session document for the user"
      operationId: "openSession"
      produces:
          - "application/json"
      responses:
        "200":
          description: "OK"
          schema:
            $ref: "#/definitions/session"
  /health:
    get:
      tags:
        - "health"
      summary: "Get health status"
      description: ""
      responses:
        "200":
          description: "The service is alive and healthy"
  /metrics:
    get:
      description: Returns metrics of the session service
      produces:
        - application/json; charset=utf-8
        - text/plain; charset=utf-8
      responses:
        '200':
          description: successful operation
          schema:
            type: array
definitions:
  session:
    description: "Describes the Qlik Associative Engine session; what id and where the session is located"
    type: "object"
    required:
      - ipAddress
      - port
      - sessionId
    properties:
      ipAddress:
        description: "IP Address of the Qlik Associative Engine"
        type: "string"
      port:
        description: "Port number exposed by the Qlik Associative Engine"
        type: "number"
      sessionId:
        description: "Guid identifying the Qlik Associative Engine session"
        type: "string"
      docId:
        description: "Guid of the doc"
        type: "string"